Fix motor timing.
Code is using half stepping, so the correct value or MICROSTEPS is 2
rather than 16.
Time delay is now measured in us, not ms.
Provide a precise delay function.
Update some of the Adafruit code with the fixes that have been applied
since it was forked.
Remove steppingcounter, which was not sufficiently precise.
Add delays to the motor test g-codes... otherwise, the motors buzz but don't turn.